Add support for Python 3
Breaking chances: the most common position is not arbitrary anymore, we
select the last one if multiple positions have the same number of
occurences. Test cases have been updated for that.

The source code for the sample user have been heavily refactored to
generate the same random records in Python 2 and 3.